% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d8flowdir.R
\name{taudem_d8flowdir}
\alias{taudem_d8flowdir}
\title{D8 Flow Directions}
\usage{
taudem_d8flowdir(
  input_elevation_grid,
  output_d8flowdir_grid = NULL,
  output_d8slopes_grid = NULL,
  n_processes = getOption("traudem.n_processes", 1)
)
}
\arguments{
\item{input_elevation_grid}{Pit filled elevation input data}

\item{output_d8flowdir_grid}{D8 flow directions output}

\item{output_d8slopes_grid}{D8 slopes output}

\item{n_processes}{Number of processes}
}
\value{
List with the two output filenames
}
\description{
D8 Flow Directions
}
\details{
See \url{https://hydrology.usu.edu/taudem/taudem5/help53/D8FlowDirections.html}
}
\examples{
\dontrun{
test_dir <- withr::local_tempdir()
 file.copy(
   system.file("test-data", "MED_01_01.tif", package = "traudem"),
   file.path(test_dir, "MED_01_01.tif")
 )
filled_pit <- taudem_pitremove(file.path(test_dir, "MED_01_01.tif"))
outputs <- taudem_d8flowdir(filled_pit)
outputs
}
}
