% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taudem-exec.R
\name{taudem_exec}
\alias{taudem_exec}
\title{Call TauDEM}
\usage{
taudem_exec(n_processes, args)
}
\arguments{
\item{n_processes}{Number of processes for \code{mpiexec}. If \code{NULL} TauDEM is called without mpiexec.}

\item{args}{Character vector of argument, starting with the TauDEM command. See examples.}
}
\value{
None
}
\description{
Call TauDEM
}
\details{
You can use this function to call more TauDEM methods
than the ones with dedicated wrappers in this package.
}
\examples{
\dontrun{
test_dir <- withr::local_tempdir()
dir.create(test_dir)
 file.copy(
   system.file("test-data", "MED_01_01.tif", package = "traudem"),
   file.path(test_dir, "MED_01_01.tif")
 )
taudem_exec(n_processes = NULL, args = c("pitremove", file.path(test_dir, "MED_01_01.tif")))
}
}
