% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{taudem_sitrep}
\alias{taudem_sitrep}
\title{TauDEM situation report}
\usage{
taudem_sitrep()
}
\value{
None
}
\description{
Checks installation of TauDEM and provides useful hints.
}
\section{TauDEM installation and registration}{


Once you have installed TauDEM add an environment variable
pointing to the correct path.
For instance on Ubuntu it could be adding this line in \code{.Renviron}
(see \code{usethis::edit_r_environ()})
and then re-starting R:

\if{html}{\out{<div class="sourceCode">}}\preformatted{TAUDEM_PATH='/usr/local/taudem'
}\if{html}{\out{</div>}}

or, for just the session, running this line of R code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv(TAUDEM_PATH='/usr/local/taudem')
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
try(taudem_sitrep(), silent = TRUE)
}
}
