% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traudem-package.R
\docType{package}
\name{traudem-package}
\alias{traudem}
\alias{traudem-package}
\title{traudem: Use 'TauDEM'}
\description{
Simple trustworthy utility functions to use 'TauDEM' command-line interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucarraro.github.io/traudem/}
  \item \url{https://github.com/lucarraro/traudem}
  \item Report bugs at \url{https://github.com/lucarraro/traudem/issues}
}

}
\author{
\strong{Maintainer}: Luca Carraro \email{Luca.Carraro@eawag.ch}

Authors:
\itemize{
  \item Maëlle Salmon \email{maelle@cynkra.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
  \item Wael Sadek
}

Other contributors:
\itemize{
  \item University of Zurich [copyright holder, funder]
}

}
\keyword{internal}
