% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitremove.R
\name{taudem_pitremove}
\alias{taudem_pitremove}
\title{Pit Remove}
\usage{
taudem_pitremove(
  input_elevation_grid,
  output_elevation_grid = NULL,
  only_4way_neighbors = FALSE,
  n_processes = getOption("traudem.n_processes", 1),
  depmask = NULL
)
}
\arguments{
\item{input_elevation_grid}{Input elevation grid file}

\item{output_elevation_grid}{Output elevation grid file}

\item{only_4way_neighbors}{Whether to consider only 4 way neighbors}

\item{n_processes}{Number of processes}

\item{depmask}{Depression mask file (optional)}
}
\value{
Path to output file (invisibly)
}
\description{
Pit Remove
}
\details{
See https://hydrology.usu.edu/taudem/taudem5/help53/PitRemove.html
}
\examples{
\dontrun{
test_dir <- withr::local_tempdir()
 fs::file_copy(
   system.file("test-data", "MED_01_01.tif", package = "traudem"),
   file.path(test_dir, "MED_01_01.tif")
 )
output <- taudem_pitremove(file.path(test_dir, "MED_01_01.tif"))
output
}
}
